chore: reorganize score UI packages
